+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # ChangeLog
 
 
+## 0.0.65 (April 8, 2018)
+- Cyan-ed embedded section tags in JSX/TSX.
+- Added additional JSX samples.
+
 ## 0.0.64 (April 7, 2018)
 - Fixed highlighting for quoted keys in TS.